reusable code moved to util
